Prevent duplicate security creation
Since securities are created AND THEN added to the algorithm, it was
possible for two different universes to create the SAME security at
the same time step. This led to two different Security instances and
only one of them being added. Ironically, it appears that it was a last
in wins situation.

This change aims to track securities that are created by a universe in a
given time step and re-use the already created security. A more holistic
solution here may be to encapsulate the creation/fetching of securities
into a new type and then ensure we use that abstraction everywhere, but
this would have led to many more changes.