Performance speedup in SubscriptionCollection
The list of subscriptions needs to be sorted by TickType when enumerating, but this sorting was previously done in GetEnumerator (called for each data point).
This has been fixed by sorting only when adding or removing subscriptions, gaining a significant backtest speed improvement.

In addition, the storage for subscriptions was simplified to a single ConcurrentDictionary, instead of nested dictionaries.