Configurable Composer dll look-up directory
When creating a collection of directories from where dynamic-link library can be loaded from, a user-defined path is used if the reserved key `composer-dll-directory` exists. Otherwise, the current executing directory is used.
This feature is necessary for applications, e.g.: Jupyter kernels, that use Lean/QuantConnect assemblies, since they may be called from a directory with dlls that don't need to be loaded by the `Composer`.