Hack fine universe data into the algorithm
Fine universes are considered special due to their chaining behavior with
the coarse universe. UniverseSelection.ApplyUniverseSelection manually
reads the fine data for symbols that have passed the initial coarse filter.
There doesn't exist a mechanism for pipe this data back to the data feed
for inclusion in the TimeSlice object. As such, this hack relies on reference
semantics to updates the BaseDataCollection to include the merged coarse and
fine fundamental data.