Implements CallsOnly and PutsOnly methods in OptionFilterUniverse
These two helper methods makes it simpler to select options with a specific right. This selection could be done with a Linq expression, but there is no equivalent for python. It is more user friendly to add those helper than adding a `PyObject` overload to `Contracts` method.

Use `PutsOnly` method in `BasicTemplateOptionsFrameworkAlgorithm`.