Fix empty dataframe from QuantBook.Indicator
When a history request is made to update a indicator, it can return an empty enumerable if there is no data from the default data type (e.g.: `QuoteBar` for Crypto). In this case, the indicator should be updated with the available data.

We changed the History requests within `QuantBook.Indicator` the  to fetch `IEnumerable<Slice>` and assure that any data is present.