Fix IB Tick History Requests Returning Second Bars
Currently tick resolution history requests are returning second bars instead of returning ticks.

IB added a new reqHistoricalTicks function for tick historical data in a recent upgrade. We recently upgraded the IB Gateway client in the cloud to the latest version, so we can now implement historical tick requests properly.

In this PR, for now we return an empty result for Tick history requests as returning second bars instead causes a runtime error (as reported in #1837).